refactor: centralize readline utilities to fix Jest open handle issues
- Create new src/util/readline.ts module with centralized readline functions

- Extract createReadlineInterface(), promptUser(), and promptYesNo() utilities

- Update src/util/server.ts, src/feedback.ts, and src/evaluator.ts to use new utilities

- Remove duplicate readline interface creation code

- Fix Jest open handle warnings by properly mocking readline utilities

- Remove redundant promptConfirm function (was just wrapper around promptYesNo)

This refactoring improves code reuse, testability, and eliminates TTYWRAP open handle issues in Jest tests.
